chore: moved industryArticle to comp
