+++ scrybe_frontend/src/components/Industry/industryArticle.jsx
